# Conflicts:
#	src/main/java/web/controller/UserController.java

+++ src/main/java/web/controller/UserController.java
@@ -17,13 +17,9 @@ import web.service.UserService;
 @RequestMapping("/")
 public class UserController {
 
-
+    @Autowired
     private UserService userService;
 
-    @Autowired
-    public UserController(UserService userService) {
-        this.userService = userService;
-    }
 
     @GetMapping("users")
     public String listUser(Model model) {
@@ -43,10 +39,17 @@ public class UserController {
         return "redirect:/users";
     }
 
-    @GetMapping("/delete/{id}")
+    @DeleteMapping("/delete/{id}")
     public String deleteUser(@PathVariable("id") Long id) {
-        userService.removeUser(id);
-        return "redirect:/users";
+        try {
+            User user = userService.getUserById(id);
+            if(user != null) {
+                userService.removeUser(id);
+            }
+            return "redirect:/users";
+        } catch (Exception e) {
+            return "user not found";
+        }
     }
 
     @GetMapping("/update/{id}")
@@ -61,4 +64,5 @@ public class UserController {
         userService.updateUser(user);
         return "redirect:/users";
     }
+
 }
